==> 0-2-data-merge.ipynb <==
import pandas as pd

Merge text

path_train = 'data/arguments-training.tsv'

path_val = 'data/arguments-validation.tsv'
path_val_zhihu = 'data/arguments-validation-zhihu.tsv'

path_test = 'data/arguments-test.tsv'
path_test_nahj = 'data/arguments-test-nahjalbalagha.tsv'
path_test_nyt = 'data/arguments-test-nyt.tsv'

dftrain = pd.read_csv(path_train, index_col=0, delimiter='\t')

dfval = pd.read_csv(path_val, index_col=0, delimiter='\t')
dfvalzhihu = pd.read_csv(path_val_zhihu, index_col=0, delimiter='\t')

dftest = pd.read_csv(path_test, index_col=0, delimiter='\t')
dftestnahj = pd.read_csv(path_test_nahj, index_col=0, delimiter='\t')
dftestnyt = pd.read_csv(path_test_nyt, index_col=0, delimiter='\t')

dftrain['Type'] = 'train'

dfval['Type'] = 'val'
dfvalzhihu['Type'] = 'zhihu'

dftest['Type'] = 'test'
dftestnahj['Type'] = 'nahj'
dftestnyt['Type'] = 'nyt'

df = pd.concat([dftrain, dfval, dfvalzhihu, dftest, dftestnahj, dftestnyt])

df.to_csv('data/text.tsv', sep='\t')

Merge labels 20

path_train = 'data/labels-training.tsv'

path_val = 'data/labels-validation.tsv'
path_val_zhihu = 'data/labels-validation-zhihu.tsv'

dftrain = pd.read_csv(path_train, index_col=0, delimiter='\t')

dfval = pd.read_csv(path_val, index_col=0, delimiter='\t')
dfvalzhihu = pd.read_csv(path_val_zhihu, index_col=0, delimiter='\t')

dftrain['Type'] = 'train'

dfval['Type'] = 'val'
dfvalzhihu['Type'] = 'zhihu'

df = pd.concat([dftrain, dfval, dfvalzhihu])

df.to_csv('data/labels.tsv', sep='\t')

Merge labels 54

path_train = 'data/level1-labels-training.tsv'

path_val = 'data/level1-labels-validation.tsv'
path_val_zhihu = 'data/level1-labels-validation-zhihu.tsv'

dftrain = pd.read_csv(path_train, index_col=0, delimiter='\t')

dfval = pd.read_csv(path_val, index_col=0, delimiter='\t')
dfvalzhihu = pd.read_csv(path_val_zhihu, index_col=0, delimiter='\t')

dftrain['Type'] = 'train'

dfval['Type'] = 'val'
dfvalzhihu['Type'] = 'zhihu'

df = pd.concat([dftrain, dfval, dfvalzhihu])

df.to_csv('data/level1-labels.tsv', sep='\t')
