==> 1-2-data-prepare-zshot.ipynb <==
import pandas as pd

text = pd.read_csv('data/text.tsv', index_col=0, delimiter='\t')
labels = pd.read_csv('data/labels.tsv', index_col=0, delimiter='\t')
level1labels = pd.read_csv('data/level1-labels.tsv', index_col=0, delimiter='\t')

df20 = pd.merge(text, labels, left_index=True, right_index=True).drop('Type_y', axis=1)
df54 = pd.merge(text, level1labels, left_index=True, right_index=True).drop('Type_y', axis=1)

df20.loc[df20['Type_x']=='train'].to_csv('data/prepared/train-zshot-20.tsv', index=False, sep='\t')
df54.loc[df54['Type_x']=='train'].to_csv('data/prepared/train-zshot-54.tsv', index=False, sep='\t')

len(labels)
