==> 2-2-model-mlabel.ipynb <==
import torch
import numpy as np
import pandas as pd
from datasets import load_dataset
from transformers import AutoTokenizer, DataCollatorWithPadding, TrainingArguments, set_seed
from transformers import AutoModelForSequenceClassification, Trainer
from sklearn.metrics import f1_score, classification_report
import os

SEED = 2387
set_seed(SEED)
import logging
#logging.getLogger("transformers").setLevel(logging.ERROR)

# Params & Settings

ex = 23
num_epochs = 5
checkpoint = 'bert-base-uncased'

# Helper Funcs

def sigmoid(X):
    return 1 / (1 + np.exp(-X))

def heaviside(X):
    return np.heaviside(X - 0.5, 0)

def f1(preds):
    logits, y_true = preds
    y_pred = heaviside(sigmoid(logits))
    return {'f1':f1_score(y_true, y_pred, average='macro')}

# Loading datasets (train, val, zhihu)

custom_tokens = pd.read_csv(f'experiments/ex{ex}/custom-tokens-mlabel.csv')

# custom_tokens = pd.read_csv(f'experiments/ex{ex}/custom-tokens-mlabel.csv')
# custom_tokens = custom_tokens['0'].tolist()
custom_tokens = []

#ds_url = 'https://zenodo.org/record/7550385/files/'
ds_url = f'experiments/ex{ex}/'
ds_files = {
    'train': ds_url + 'train-mlabel.tsv',
    'val': ds_url + 'val-mlabel.tsv',
    'val.zhihu': ds_url + 'zhihu-mlabel.tsv'
}

ds = load_dataset('csv', data_files=ds_files, delimiter='\t')
ds = ds.rename_columns({'Text':'text', 'Labels':'label', 'Argument ID': 'id'})
ds

import ast
def convert_labels(example):
    example["label"] = [float(i) for i in ast.literal_eval(example["label"])]
    return example

ds = ds.map(convert_labels)


# Loading Model & Tokenizer & Collator

custom_tokens

tokenizer = AutoTokenizer.from_pretrained(checkpoint)

# tokenizer.add_tokens(custom_tokens)
# vocab_size = tokenizer.vocab_size

def tokenize_function(example):
    return tokenizer(example['text'], truncation=True)


tokenized_datasets = ds.map(tokenize_function, batched=True)
data_collator = DataCollatorWithPadding(tokenizer=tokenizer)

training_args = TrainingArguments(
    f'experiments/ex{ex}/models', True,
    num_train_epochs=num_epochs,
    
    evaluation_strategy='epoch',
    save_strategy='epoch'
)
model = AutoModelForSequenceClassification.from_pretrained(checkpoint, num_labels=20, problem_type="multi_label_classification")
# model.resize_token_embeddings(vocab_size)

trainer = Trainer(
    model,
    training_args,
    train_dataset=tokenized_datasets['train'],
    eval_dataset=tokenized_datasets['val'],
    data_collator=data_collator,
    tokenizer=tokenizer,
    compute_metrics=f1,
)

# Training the Model

train_output = trainer.train()

trainable_params = sum(p.numel() for p in model.parameters() if p.requires_grad)
device_name = torch.cuda.get_device_name(torch.device('cuda'))

repline = f'{ex},{SEED},{checkpoint},{trainable_params},{train_output[2]["train_runtime"]},{device_name},{len(tokenized_datasets["train"])},{num_epochs}'

res_train = trainer.evaluate(tokenized_datasets['train'])
res_val = trainer.evaluate(tokenized_datasets['val'])
res_zhihu = trainer.evaluate(tokenized_datasets['val.zhihu'])

repline += f',{res_train["eval_f1"]},{res_val["eval_f1"]},{res_zhihu["eval_f1"]}'
repline += f',{res_train["eval_loss"]},{res_val["eval_loss"]},{res_zhihu["eval_loss"]}'

repline

with open('experiments/report.csv', 'a') as file:
    file.write(f'{repline}\n')

preds_train = trainer.predict(tokenized_datasets['train'])
preds_val = trainer.predict(tokenized_datasets['val'])
preds_zhihu = trainer.predict(tokenized_datasets['val.zhihu'])

res_folder = f'experiments/ex{ex}/results'
os.mkdir(res_folder)

pd.DataFrame(preds_train[0]).to_csv(f'{res_folder}/train_pred.csv', index=False)
pd.DataFrame(preds_train[1]).to_csv(f'{res_folder}/train_true.csv', index=False)

pd.DataFrame(preds_val[0]).to_csv(f'{res_folder}/val_pred.csv', index=False)
pd.DataFrame(preds_val[1]).to_csv(f'{res_folder}/val_true.csv', index=False)

pd.DataFrame(preds_zhihu[0]).to_csv(f'{res_folder}/zhihu_pred.csv', index=False)
pd.DataFrame(preds_zhihu[1]).to_csv(f'{res_folder}/zhihu_true.csv', index=False)

#ds_url = 'https://zenodo.org/record/7550385/files/'
ds_url = f'experiments/ex{ex}/'
ds_files = {
    'test': ds_url + 'test-mlabel.tsv',
    'test.nahj': ds_url + 'nahj-mlabel.tsv',
    'test.nyt': ds_url + 'nyt-mlabel.tsv'
}

ds = load_dataset('csv', data_files=ds_files, delimiter='\t')
ds = ds.rename_columns({'Text':'text', 'Argument ID': 'id'})
ds

tokenizer = AutoTokenizer.from_pretrained(checkpoint)

# tokenizer.add_tokens(custom_tokens)
# vocab_size = tokenizer.vocab_size

def tokenize_function(example):
    return tokenizer(example['text'], truncation=True)


tokenized_datasets = ds.map(tokenize_function, batched=True)
data_collator = DataCollatorWithPadding(tokenizer=tokenizer)

preds_test = trainer.predict(tokenized_datasets['test'])
preds_nahj = trainer.predict(tokenized_datasets['test.nahj'])
preds_nyt = trainer.predict(tokenized_datasets['test.nyt'])

cols = ['Self-direction: thought','Self-direction: action','Stimulation','Hedonism','Achievement','Power: dominance','Power: resources','Face','Security: personal','Security: societal','Tradition','Conformity: rules','Conformity: interpersonal','Humility','Benevolence: caring','Benevolence: dependability','Universalism: concern','Universalism: nature','Universalism: tolerance','Universalism: objectivity']

pd.DataFrame(heaviside(sigmoid(preds_test[0])), index=ds['test']['id'], columns=cols).astype('int').to_csv(f'experiments/ex{ex}/BERT-test.tsv', sep='\t')
pd.DataFrame(heaviside(sigmoid(preds_nahj[0])), index=ds['test.nahj']['id'], columns=cols).astype('int').to_csv(f'experiments/ex{ex}/BERT-nahj.tsv', sep='\t')
pd.DataFrame(heaviside(sigmoid(preds_nyt[0])), index=ds['test.nyt']['id'], columns=cols).astype('int').to_csv(f'experiments/ex{ex}/BERT-nyt.tsv', sep='\t')
