==> 2-3-model-ensemble-ex4-ex5.ipynb <==
import pandas as pd
import numpy as np
from sklearn.metrics import f1_score, classification_report

def sigmoid(X):
    return 1 / (1 + np.exp(-X))

def heaviside(X):
    return np.heaviside(X - 0.5, 0)

preds_train_4 = np.genfromtxt('experiments/ex4/results/train_pred.csv', delimiter=',', skip_header=1)
true_train_4 = np.genfromtxt('experiments/ex4/results/train_true.csv', delimiter=',', skip_header=1)
preds_val_4 = np.genfromtxt('experiments/ex4/results/val_pred.csv', delimiter=',', skip_header=1)
true_val_4 = np.genfromtxt('experiments/ex4/results/val_true.csv', delimiter=',', skip_header=1)
preds_zhihu_4 = np.genfromtxt('experiments/ex4/results/zhihu_pred.csv', delimiter=',', skip_header=1)
true_zhihu_4 = np.genfromtxt('experiments/ex4/results/zhihu_true.csv', delimiter=',', skip_header=1)

preds_train_5 = np.genfromtxt('experiments/ex5/results/train_pred.csv', delimiter=',', skip_header=1)
true_train_5 = np.genfromtxt('experiments/ex5/results/train_true.csv', delimiter=',', skip_header=1)
preds_val_5 = np.genfromtxt('experiments/ex5/results/val_pred.csv', delimiter=',', skip_header=1)
true_val_5 = np.genfromtxt('experiments/ex5/results/val_true.csv', delimiter=',', skip_header=1)
preds_zhihu_5 = np.genfromtxt('experiments/ex5/results/zhihu_pred.csv', delimiter=',', skip_header=1)
true_zhihu_5 = np.genfromtxt('experiments/ex5/results/zhihu_true.csv', delimiter=',', skip_header=1)

preds_train = heaviside(sigmoid(np.concatenate((preds_train_4, preds_train_5))))
true_train = np.concatenate((true_train_4, true_train_5))
preds_val = heaviside(sigmoid(np.concatenate((preds_val_4, preds_val_5))))
true_val = np.concatenate((true_val_4, true_val_5))
preds_zhihu = heaviside(sigmoid(np.concatenate((preds_zhihu_4, preds_zhihu_5))))
true_zhihu = np.concatenate((true_zhihu_4, true_zhihu_5))

train_f1 = f1_score(true_train, preds_train, average='macro')
val_f1 = f1_score(true_val, preds_val, average='macro')
zhihu_f1 = f1_score(true_zhihu, preds_zhihu, average='macro')

[train_f1, val_f1, zhihu_f1]
